+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update npm dependencies
 - Multiple accessibility improvements
 - Upgrade to `node:8.12.0-alpine`
+- Fix skip to content for voice over
 
 0.10.0 / 2018-09-14
 ==================